Use the s390x release image
This is a hack. We are using the AMD64 build of the cli tools and then
swapping out the release image for the s390x one.

This isn't a CI build, but is instead a nightly. This should be okay for
now, but we'll want to update this once we get the full s390x build farm
going.